add changes to my page and css
