+++ src/components/Search.css
@@ -27,7 +27,7 @@ header {
   flex-direction: column;
   align-items: center; /* Center items horizontally */
   justify-content: center; /* Center items vertically */
-  min-height: 500px;
+  min-height: 100vh;
   height: 100%; /* Full viewport height */
   background-color: #2c2b4a;
   padding: 20px;
@@ -50,12 +50,18 @@ header {
   display: flex;
   align-items: center;
   justify-content: space-between;
-  width: 300px; /* Adjusted width for consistency */
+  width: 100%;
   margin-bottom: 10px;
   padding: 10px;
-  background-color: #fff; /* White background for better readability */
+  background-color: #e7e3d4; /* White background for better readability */
   border-radius: 5px; /* Rounded corners */
   box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1); /* Subtle shadow */
+}
+
+@media (min-width: 768px) {
+  .user-item {
+    max-width: 40%; /* Змінюємо максимальну ширину на 40% для екранів ноутбуків і більших */
+  }
 }
 
 /* Button container styling */
